requirements.txt
# ============================================================================
# PlotPointe - Multi-Modal GAT Recommendation System
# Production Dependencies
# Python 3.10+ compatible
# ============================================================================

# Deep Learning Framework (MUST install first - pinned versions from debugging)
# These versions are tested and working on GCP Vertex AI containers
torch==2.1.0
torchvision==0.16.0

# Transformers & Embeddings (pinned to compatible versions)
# transformers 4.35.2 is compatible with torch 2.1.0
transformers==4.35.2
sentence-transformers>=2.2.0,<3.0.0

# Google Cloud Platform
# Vertex AI SDK for experiment tracking and custom jobs
google-cloud-aiplatform>=1.35.0,<2.0.0
google-cloud-storage>=2.10.0,<3.0.0
google-cloud-bigquery>=3.11.0,<4.0.0

# Data Processing
# pandas 2.x for better performance and type hints
pandas>=2.0.0,<3.0.0
pyarrow>=12.0.0,<15.0.0
numpy>=1.24.0,<2.0.0  # numpy 2.x has breaking changes

# Scientific Computing
scipy>=1.10.0,<2.0.0

# Machine Learning
scikit-learn>=1.3.0,<2.0.0

# Image Processing
# Pillow for image loading and preprocessing
pillow>=10.0.0,<11.0.0

# Utilities
requests>=2.31.0,<3.0.0
tqdm>=4.65.0,<5.0.0

# Graph Processing
# networkx for graph algorithms and analysis
networkx>=3.1,<4.0.0

# OPTIONAL: Fast similarity search (10-100x faster than sklearn)
# Uncomment for production deployment with large-scale kNN
# faiss-cpu>=1.7.4,<2.0.0  # CPU-only version
# faiss-gpu>=1.7.4,<2.0.0  # GPU version (requires CUDA)

